requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.95.1
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0

# JWT 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1

# STT (음성 인식)
speechrecognition==3.10.0
pyaudio==0.2.11

# 환경 변수
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# 이메일 검증
email-validator==2.1.0
